Addd final documentations.
